refactor(orchestrator): extract query and timeout operations from GoldenRecordTaskService
